use 64-bit block offsets instead of capping positions per block

+++ src/Storages/MergeTree/TextIndexBlockedPositionsCodec.h
@@ -59,7 +59,7 @@ public:
     struct DecodeScratch
     {
         PaddedPODArray<char> payload;
-        PaddedPODArray<UInt32> doc_offsets;
+        PaddedPODArray<UInt64> doc_offsets;
         PaddedPODArray<UInt32> values;
     };
